refac: move main out and separate logic for model and repo
